Alteration of Equip Commands.
Whilst using this code for my own project I found the Equip Commands seemed to not make use of the cmd attribute of commands which indicates the alias used. Instead it had a master equip command and separate commands for aliases which simply started the equip command with additional arguments to indicate which command was used. This code achieves the same thing but in a single command through the use of a reference dictionary. 

Additionally, when an argument isn't passed or the command fails due to the object, a list of available items in your inventory is displayed for the command. I.e. a failed "wield" command would display a list of available wieldable items in your inventory as opposed to your functionality of displaying your equipped items. My usage involved showing a list of equipped items in another command, but if you would prefer the original functionality of displaying that list when using a equip command without arguments let me know and I'll add that.